// src/interfaces.tsx
export interface City {
    ACCESS2_Adj95CI?: string;
    ACCESS2_AdjPrev?: string;
    ACCESS2_Crude95CI?: string;
    ACCESS2_CrudePrev?: string;
    ARTHRITIS_Adj95CI?: string;
    ARTHRITIS_AdjPrev?: string;
    ARTHRITIS_Crude95CI?: string;
    ARTHRITIS_CrudePrev?: string;
    BINGE_Adj95CI?: string;
    BINGE_AdjPrev?: string;
    BINGE_Crude95CI?: string;
    BINGE_CrudePrev?: string;
    BPHIGH_Adj95CI?: string;
    BPHIGH_AdjPrev?: string;
    BPHIGH_Crude95CI?: string;
    BPHIGH_CrudePrev?: string;
    BPMED_Adj95CI?: string;
    BPMED_AdjPrev?: string;
    BPMED_Crude95CI?: string;
    BPMED_CrudePrev?: string;
    CANCER_Adj95CI?: string;
    CANCER_AdjPrev?: string;
    CANCER_Crude95CI?: string;
    CANCER_CrudePrev?: string;
    CASTHMA_Adj95CI?: string;
    CASTHMA_AdjPrev?: string;
    CASTHMA_Crude95CI?: string;
    CASTHMA_CrudePrev?: string;
    CHD_Adj95CI?: string;
    CHD_AdjPrev?: string;
    CHD_Crude95CI?: string;
    CHD_CrudePrev?: string;
    CHECKUP_Adj95CI?: string;
    CHECKUP_AdjPrev?: string;
    CHECKUP_Crude95CI?: string;
    CHECKUP_CrudePrev?: string;
    CHOLSCREEN_Adj95CI?: string;
    CHOLSCREEN_AdjPrev?: string;
    CHOLSCREEN_Crude95CI?: string;
    CHOLSCREEN_CrudePrev?: string;
    COLON_SCREEN_Adj95CI?: string;
    COLON_SCREEN_AdjPrev?: string;
    COLON_SCREEN_Crude95CI?: string;
    COLON_SCREEN_CrudePrev?: string;
    COPD_Adj95CI?: string;
    COPD_AdjPrev?: string;
    COPD_Crude95CI?: string;
    COPD_CrudePrev?: string;
    COREM_Adj95CI?: string;
    COREM_AdjPrev?: string;
    COREM_Crude95CI?: string;
    COREM_CrudePrev?: string;
    COREW_Adj95CI?: string;
    COREW_AdjPrev?: string;
    COREW_Crude95CI?: string;
    COREW_CrudePrev?: string;
    CSMOKING_Adj95CI?: string;
    CSMOKING_AdjPrev?: string;
    CSMOKING_Crude95CI?: string;
    CSMOKING_CrudePrev?: string;
    DENTAL_Adj95CI?: string;
    DENTAL_AdjPrev?: string;
    DENTAL_Crude95CI?: string;
    DENTAL_CrudePrev?: string;
    DIABETES_Adj95CI?: string;
    DIABETES_AdjPrev?: string;
    DIABETES_Crude95CI?: string;
    DIABETES_CrudePrev?: string;
    Geolocation: string;
    HIGHCHOL_Adj95CI?: string;
    HIGHCHOL_AdjPrev?: string;
    HIGHCHOL_Crude95CI?: string;
    HIGHCHOL_CrudePrev?: string;
    KIDNEY_Adj95CI?: string;
    KIDNEY_AdjPrev?: string;
    KIDNEY_Crude95CI?: string;
    KIDNEY_CrudePrev?: string;
    LPA_Adj95CI?: string;
    LPA_AdjPrev?: string;
    LPA_Crude95CI?: string;
    LPA_CrudePrev?: string;
    MAMMOUSE_Adj95CI?: string;
    MAMMOUSE_AdjPrev?: string;
    MAMMOUSE_Crude95CI?: string;
    MAMMOUSE_CrudePrev?: string;
    MHLTH_Adj95CI?: string;
    MHLTH_AdjPrev?: string;
    MHLTH_Crude95CI?: string;
    MHLTH_CrudePrev?: string;
    OBESITY_Adj95CI?: string;
    OBESITY_AdjPrev?: string;
    OBESITY_Crude95CI?: string;
    OBESITY_CrudePrev?: string;
    PAPTEST_Adj95CI?: string;
    PAPTEST_AdjPrev?: string;
    PAPTEST_Crude95CI?: string;
    PAPTEST_CrudePrev?: string;
    PHLTH_Adj95CI?: string;
    PHLTH_AdjPrev?: string;
    PHLTH_Crude95CI?: string;
    PHLTH_CrudePrev?: string;
    PlaceFIPS?: string;
    PlaceName: string;
    Population2010?: string;
    SLEEP_Adj95CI?: string;
    SLEEP_AdjPrev?: string;
    SLEEP_Crude95CI?: string;
    SLEEP_CrudePrev?: string;
    STROKE_Adj95CI?: string;
    STROKE_AdjPrev?: string;
    STROKE_Crude95CI?: string;
    STROKE_CrudePrev?: string;
    StateAbbr?: string;
    TEETHLOST_Adj95CI?: string;
    TEETHLOST_AdjPrev?: string;
    TEETHLOST_Crude95CI?: string;
    TEETHLOST_CrudePrev?: string;
};

export interface CityProcessed {
    stateAbbr: string;
    placeName: string;
    lng: number;
    lat: number;
    lnglat: [number, number];
}

export interface MapTransitionState {
    cities?: City[];
    usmap?: object;
};

export interface MapTransitionProps {
    width: number;
    height: number;
};